catalog-service/src/main/java/com/adnan/catalog/dto/ProductDto.java
package com.adnan.catalog.dto;

import java.time.LocalDate;

public class ProductDto {
    
    private Long id;
    private String name;
    private String description;
    private String imageUrl;
    private double price;
    private boolean inStock;
    private Long categoryId;

    public ProductDto() {
    }

    public ProductDto(String name, String description, String imageUrl, double price, boolean inStock, Long categoryId) {
        this.name = name;
        this.description = description;
        this.imageUrl = imageUrl;
        this.price = price;
        this.inStock = inStock;
        this.categoryId = categoryId;
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getImageUrl() {
        return imageUrl;
    }

    public void setImageUrl(String imageUrl) {
        this.imageUrl = imageUrl;
    }

    public double getPrice() {
        return price;
    }

    public void setPrice(double price) {
        this.price = price;
    }

    public boolean isInStock() {
        return inStock;
    }

    public void setInStock(boolean inStock) {
        this.inStock = inStock;
    }

    public Long getCategoryId() {
        return categoryId;
    }

    public void setCategoryId(Long categoryId) {
        this.categoryId = categoryId;
    }

}
